Refactor KV cache, Rope  , reduce common code  (#1148)
